setproctitle [was: Re: Removal of the 6.x kernel compat code from libc]
On Fri, Apr 17, 2015 at 01:39:04PM +0200, Oliver Pinter wrote:
> Is there any chanche to get ride of the very old (FreeBSD 2.x) compat
> hacks like these:
> https://github.com/freebsd/freebsd/blob/master/lib/libc/gen/setproctitle.c#L40 ?
>

Below is the promised cleanup of the libc/gen/setproctitle.c.

I see a value in keeping the track of the historical interfaces. I
considered moving the explanation about old_ps_strings to sys/exec.h,
where the current ps_strings is defined and explained, but this would
make the safety checks in the setproctitle() code uncomprehensive.

CC: Oliver Pinter <op@hardenedbsd.org>