HBSD: Add knob for CFI
The Control-Flow Integrity project from clang/llvm provides additional
security at the cost of an overhead between 1% and 12%. As of this
time, CFI cannot be enabled for all of base.

This commit provides an opt-in framework, similar to the initial
approach to PIE. Users will be able to specify WITH_CFI in src.conf(5),
and those applications in base that claim they are safe to compile with
CFI will have CFI enabled.

This only applies CFI to executables, not to DSOs. Applying CFI to DSOs
will require a lot of research and development. As such, the CFI option
to apply cross-DSO checks is not enabled.

CFI requires LLD or GNU Gold. For purposes of this commit, we require
that LLD be enabled and configured as the linker in base. A follow-up
commit will forcibly disable CFI if either MK_LLD or MK_LLD_IS_LD are
set to "no".

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX