HBSD: Add ASAN support
ASAN is a good development and testing tool to ensure sane memory
access. It is not meant to be used as an exploit mitigation, but rather
a tool to be used during the development lifecycle. As such, do not
enable it by default in HardenedBSD, but provide the ability to make use
of it for downstream consumers during the development lifecycle.

To use ASAN, set WITH_ASAN in src.conf(5).

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
GitHub-issue:	#260