HBSD: bhyve obfuscation, part 1 of many
Give users the ability to obfuscate away the bhyve identification
string.

This is just one commit in a series of commits that will enhance bhyve
with the ability to obfuscate the identification of the hypervisor. This
is especially useful when running untrusted code in untrusted virtual
machines (ie, malware).

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	G2, Inc
(cherry picked from commit a52c62a5ed5d375b816482b5be6771ecdeaa4a02)