HBSD ASAN: Lower entropy introduced into the stack
llvm's ASAN implementation does not like how aggressive we are with
stack randomization. After toying around with the stack and VDSO deltas,
I've found that 18 bits of entropy for the stack and 10 for the VDSO
seem to be the most robust with ASAN.

Additionally, llvm's documentation states that ASAN can cause up to
three times the normal stack utilization. Reflect that in the max stack
size by bumping it up to 3GB from its original 1GB.

Note that this weakens the ASLR implementation greatly and this commit
should not be merged or cherry-picked into any other branch.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX