HBSD: Disable Cross-DSO CFI for jevents
jevents is used to autogenerate C code during the build process. It
segfaults when Cross-DSO CFI is enabled. So disable CFI, but keep LTO
since the rest of the world is compiled with LTO.

The intention is to come back to this application later. However, libpmc
is seeing a lot of code churn in upstream FreeBSD right now. It's best
to wait till upstream settles down and I get Cross-DSO CFI working with
statically-compiled applications.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>