HBSD: Revert "Convert amd64_get/set_fs/gsbase to ifunc."
This reverts commit 727359fd37194926818b81e49aa594cc03a9aeb9. This
causes a lot of breakages in HardenedBSD. Base breaks due to our strict
usage of full RELRO (RELRO + BIND_NOW). Ports breaks due to our use of
full RELRO in addition to `-fstack-protector-all`.

I plan to discuss this privately with the developer of the offending
commit from upstream FreeBSD. Hopefully, this reversion is only
temporary while we come up with a proper fix.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>