HBSD: Provide workarounds for ICMP and IPv4 frag bugs
FreeBSD is vulnerable to the XNU ICMP bug due to  the
net.inet.icmp.quotelen sysctl node being increased from 8 to 548. The
author of the referenced report below suggests decreasing the ICMP quote
length back down to 8 until a real fix is provided.

In the same report, the author hints at an IPv4 reassembly issue and
suggests setting the net.inet.ip.maxfragsperpacket sysctl node to 0.
Perform this workaround as well.

This commit is only temporaryd until we at HardenedBSD get the time to
investigate and provide a real fix. Once a fix is provided, this commit
may be able to be reverted.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX
Reported-by:	https://www.reddit.com/r/BSD/comments/9v6xwg/remotely_triggerable_icmp_buffer_underwrite_in/
MFC-to:		12-STABLE
MFC-to:		11-STABLE