HBSD: Add EFIRT to the HARDENEDBSD amd64 kernel
EFIRT is an essential component for UEFI-based systems now. When
performing an installation, `bsdinstall(8)` sets up nvram by calling
`efibootmgr`, which requires the kernel components of the EFI runtime
(efirt.ko) to be available. Since bsdinstall doesn't autoload efirt.ko,
efibootmgr fails, resulting in an unbootable system.

Resolve this by enabling EFIRT in the HARDENEDBSD amd64 kernel like it
is in GENERIC.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX